Migrate test_mc_phot_kern_merging to test_phot_kernels_merging.py
