+++ diffsky/experimental/kernels/tests/test_mc_phot_kernels.py
@@ -16,65 +16,6 @@ _A = [None, 0, None, None, 0, *[None] * 4]
 calc_obs_mags_galpop = vmap(phk.calc_obs_mag, in_axes=_A)
 
 
-def test_mc_phot_kern_merging(num_halos=250):
-    ran_key = jran.key(0)
-    lc_data, tcurves = tlcg._get_weighted_lc_photdata_for_unit_testing(
-        num_halos=num_halos
-    )
-    fb = 0.156
-
-    _res = mcpk._mc_phot_kern_merging(
-        ran_key,
-        lc_data.z_obs,
-        lc_data.t_obs,
-        lc_data.mah_params,
-        lc_data.ssp_data,
-        lc_data.precomputed_ssp_mag_table,
-        lc_data.z_phot_table,
-        lc_data.wave_eff_table,
-        *dpw.DEFAULT_PARAM_COLLECTION,
-        merging_model.DEFAULT_MERGE_PARAMS,
-        DEFAULT_COSMOLOGY,
-        fb,
-        lc_data.logmp_infall,
-        lc_data.logmhost_infall,
-        lc_data.t_infall,
-        lc_data.is_central,
-        lc_data.nhalos,
-        lc_data.halo_indx,
-    )
-    phot_kern_results, phot_randoms, flux_obs, merge_prob, mstar_obs = _res
-    assert np.all(merge_prob >= 0)
-    assert np.all(merge_prob <= 1)
-    assert np.any(merge_prob > 0)
-    assert np.any(merge_prob < 1)
-
-    assert np.all(np.isfinite(mstar_obs))
-
-    obs_mags_in_plus_ex_situ = -2.5 * np.log10(flux_obs)
-    assert np.any(obs_mags_in_plus_ex_situ != phot_kern_results.obs_mags)
-
-    # Enforce centrals get brighter and satellites get dimmer
-    assert np.all(
-        obs_mags_in_plus_ex_situ[lc_data.is_central == 1]
-        <= phot_kern_results.obs_mags[lc_data.is_central == 1]
-    )
-    assert np.all(
-        obs_mags_in_plus_ex_situ[lc_data.is_central == 0]
-        >= phot_kern_results.obs_mags[lc_data.is_central == 0]
-    )
-
-    # Enforce centrals get more massive and satellites less massive
-    assert np.all(
-        mstar_obs[lc_data.is_central == 1]
-        >= 10 ** phot_kern_results.logsm_obs[lc_data.is_central == 1]
-    )
-    assert np.all(
-        mstar_obs[lc_data.is_central == 0]
-        <= 10 ** phot_kern_results.logsm_obs[lc_data.is_central == 0]
-    )
-
-
 def test_mc_specphot_kern_merging(num_halos=250):
     ran_key = jran.key(0)
     lc_data, tcurves = tlcg._get_weighted_lc_photdata_for_unit_testing(
